+++ docs/changelog.rst
@@ -15,9 +15,10 @@ v0.2 (not yet released)
 * Moved generic operations from `operations.hpp` to `algorithm.hpp`
   `#14 <https://github.com/msoeken/kitty/pull/14>`_
 
-* Print truth tables as string: ``print_binary``, ``print_raw``
+* Print truth tables as string: ``print_binary``, ``print_raw``, ``print_hex``
   `#14 <https://github.com/msoeken/kitty/pull/14>`_
   `#20 <https://github.com/msoeken/kitty/pull/20>`_
+  `#22 <https://github.com/msoeken/kitty/pull/22>`_
 
 * Added iterators to all truth tables: ``begin``, ``end``, ``rbegin``, ``rend``, ``cbegin``, ``cend``, ``crbegin``, ``crend``
   `#14 <https://github.com/msoeken/kitty/pull/14>`_